Create WebpackPluginsUtil for static classes.
