+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/TaskInstallWebpackPlugins.java
@@ -17,13 +17,8 @@ package com.vaadin.flow.server.frontend;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.UncheckedIOException;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -33,7 +28,7 @@ import elemental.json.Json;
 import elemental.json.JsonArray;
 import elemental.json.JsonObject;
 import static com.vaadin.flow.server.Constants.PACKAGE_JSON;
-import static java.nio.charset.StandardCharsets.UTF_8;
+import static com.vaadin.flow.server.frontend.WebpackPluginsUtil.PLUGIN_TARGET;
 
 /**
  * Task that installs any Flow webpack plugins into node_modules/@vaadin for use
@@ -48,8 +43,6 @@ public class TaskInstallWebpackPlugins implements FallibleCommand {
 
     private File targetFolder;
 
-    protected static final String PLUGIN_TARGET = "plugins";
-
     /**
      * Copy Flow webpack plugins into the given nodeModulesFolder.
      *
@@ -62,7 +55,7 @@ public class TaskInstallWebpackPlugins implements FallibleCommand {
 
     @Override
     public void execute() {
-        getPlugins().forEach(plugin -> {
+        WebpackPluginsUtil.getPlugins().forEach(plugin -> {
             try {
                 generatePluginFiles(plugin);
             } catch (IOException ioe) {
@@ -74,41 +67,13 @@ public class TaskInstallWebpackPlugins implements FallibleCommand {
         });
     }
 
-    /**
-     * Get names for plugins to install into node_modules.
-     *
-     * @return names of plugins to install
-     */
-    protected static List<String> getPlugins() {
-        try {
-            final JsonObject jsonFile = getJsonFile(
-                    "plugins/webpack-plugins.json");
-            if (jsonFile == null) {
-                log().error(
-                        "Couldn't locate plugins/webpack-plugins.json, no Webpack plugins for Flow will be installed."
-                                + "If webpack build fails validate flow-server jar content.");
-                return Collections.emptyList();
-            }
-
-            final JsonArray plugins = jsonFile.getArray("plugins");
-            List<String> pluginsToInstall = new ArrayList<>(plugins.length());
-            for (int i = 0; i < plugins.length(); i++) {
-                pluginsToInstall.add(plugins.getString(i));
-            }
-            return pluginsToInstall;
-        } catch (IOException ioe) {
-            throw new UncheckedIOException(
-                    "Couldn't load webpack-plugins.json file", ioe);
-        }
-    }
-
     private void generatePluginFiles(String pluginName) throws IOException {
         // Get the target folder where the plugin should be installed to
         File pluginTargetFile = new File(targetFolder, pluginName);
 
         final String pluginFolderName = PLUGIN_TARGET + "/" + pluginName + "/";
-        final JsonObject packageJson = getJsonFile(
-                pluginFolderName + PACKAGE_JSON);
+        final JsonObject packageJson = WebpackPluginsUtil
+                .getJsonFile(pluginFolderName + PACKAGE_JSON);
         if (packageJson == null) {
             log().error(
                     "Couldn't locate '{}' for plugin '{}'. Plugin will not be installed.",
@@ -136,48 +101,18 @@ public class TaskInstallWebpackPlugins implements FallibleCommand {
         final JsonArray files = packageJson.getArray("files");
         for (int i = 0; i < files.length(); i++) {
             final String file = files.getString(i);
-            FileUtils.copyURLToFile(getResourceUrl(pluginFolderName + file),
+            FileUtils.copyURLToFile(
+                    WebpackPluginsUtil.getResourceUrl(pluginFolderName + file),
                     new File(pluginTargetFile, file));
         }
         // copy package.json to plugin directory
-        FileUtils.copyURLToFile(getResourceUrl(pluginFolderName + PACKAGE_JSON),
+        FileUtils.copyURLToFile(
+                WebpackPluginsUtil
+                        .getResourceUrl(pluginFolderName + PACKAGE_JSON),
                 new File(pluginTargetFile, PACKAGE_JSON));
     }
 
-    private static JsonObject getJsonFile(String jsonFilePath)
-            throws IOException {
-        final URL urlResource = getResourceUrl(jsonFilePath);
-        if (urlResource == null) {
-            return null;
-        }
-        File jsonFile = new File(urlResource.getFile());
-        String jsonString;
-        if (!jsonFile.exists()) {
-            try (InputStream resourceAsStream = getResourceAsStream(
-                    jsonFilePath)) {
-                if (resourceAsStream != null) {
-                    jsonString = FrontendUtils.streamToString(resourceAsStream);
-                } else {
-                    return null;
-                }
-            }
-        } else {
-            jsonString = FileUtils.readFileToString(jsonFile, UTF_8);
-        }
-        return Json.parse(jsonString);
-    }
-
-    private static URL getResourceUrl(String resource) {
-        ClassLoader cl = TaskInstallWebpackPlugins.class.getClassLoader();
-        return cl.getResource(resource);
-    }
-
-    private static Logger log() {
+    private Logger log() {
         return LoggerFactory.getLogger(TaskInstallWebpackPlugins.class);
-    }
-
-    private static InputStream getResourceAsStream(String resource) {
-        ClassLoader cl = TaskInstallWebpackPlugins.class.getClassLoader();
-        return cl.getResourceAsStream(resource);
     }
 }
